requirements.txt
Django>=1.11
djangorestframework>=3.6
gunicorn>=19.7
